backend/requirements.txt
# Core Flask dependencies
Flask==2.3.2
Werkzeug>=2.3.7
click>=8.1.7
itsdangerous>=2.1.2
gunicorn==20.1.0

# Session and Security
Flask-Session==0.5.0
bcrypt==4.0.1
cryptography>=41.0.0
python-dotenv==1.0.0

# Database
mysql-connector-python==8.0.33

# CORS
Flask-Cors==3.0.10

# Monitoring and Debug
requests>=2.31.0